requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
azure-eventhub==5.15.0
python-dotenv==1.1.0

